+++ packages/manager/src/features/Longview/LongviewLanding/Gauges/Network.test.tsx
@@ -6,6 +6,8 @@ import Network, { generateUnits, generateUsedNetworkAsBytes } from './Network';
 
 afterEach(cleanup);
 
+const mockError = [{ TEXT: 'no reason', CODE: 0, SEVERITY: 3 }];
+
 const loadingStore = {
   longviewStats: {
     123: {
@@ -29,11 +31,7 @@ const errorStore = {
   longviewStats: {
     123: {
       loading: false,
-      error: [
-        {
-          reason: 'this is an error'
-        }
-      ]
+      error: mockError
     }
   }
 };
