+++ packages/manager/src/features/Longview/LongviewLanding/Gauges/RAM.test.tsx
@@ -6,6 +6,8 @@ import RAM from './RAM';
 
 afterEach(cleanup);
 
+const mockError = [{ TEXT: 'no reason', CODE: 0, SEVERITY: 3 }];
+
 const loadingStore = {
   longviewStats: {
     123: {
@@ -32,11 +34,7 @@ const errorStore = {
   longviewStats: {
     123: {
       loading: false,
-      error: [
-        {
-          reason: 'this is an error'
-        }
-      ]
+      error: mockError
     }
   }
 };
