+++ packages/manager/src/features/Longview/request.types.ts
@@ -205,6 +205,7 @@ export type AllData = LongviewCPU &
   LongviewNetwork &
   LongviewSystemInfo &
   LongviewPackages &
+  LongviewProcesses &
   Uptime &
   LongviewPortsResponse &
   LastUpdated;
